Move validation file to utils
